Various usability and logging fixes

+++ migrations/002_create_withdrawals.js
@@ -5,16 +5,16 @@ exports.up = async function (knex) {
     t.increments('id').primary();
 
     t.string('validium_tx_hash', 66).notNullable(); // bytes32 with 0x prefix
-    t.decimal('validium_block_number').notNullable();
+    t.integer('validium_block_number').notNullable();
     t.string('message_hash', 66).notNullable().unique(); // bytes32 with 0x prefix
 
     t.string('from', 42).notNullable(); // address with 0x prefix
     t.string('to', 42).notNullable(); // address with 0x prefix
     t.bigint('value').notNullable(); // uint256
-    t.bigint('nonce').notNullable(); // uint256
+    t.integer('nonce').notNullable(); // uint256, but range is much smaller
     t.text('message').notNullable(); // bytes
 
-    t.decimal('batch_index');
+    t.integer('batch_index');
     t.text('proof');
 
     t.timestamps(true, true); // created_at, updated_at
